rename global directory to container
