+++ src/container/utils/getGlobal.ts
