+++ src/container/utils/setGlobal.ts
