+++ src/initialize.ts
@@ -1,5 +1,5 @@
-import { container as defaultContainer } from "./global/container";
-import { getRouter } from "./global/utils/getRouter";
+import { container as defaultContainer } from "./container/container";
+import { getRouter } from "./container/utils/getRouter";
 
 export const initialize = (container = defaultContainer) => {
   return getRouter(container).handle;
